+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/Storage.java
@@ -176,11 +176,13 @@ public abstract class Storage {
    * Runs a function on each pair of non-missing elements in this and arg.
    *
    * @param name a name of potential vectorized variant of the function that should be used if
-   *     supported. If this argument is null, the vectorized operation will never be used. *
+   *     supported. If this argument is null, the vectorized operation will never be used.
    * @param function the function to run.
+   * @param skipNa whether rows containing missing values should be passed to the function.
    * @return the result of running the function on all non-missing elements.
    */
-  public final Storage zip(String name, BiFunction<Object, Object, Object> function, Storage arg) {
+  public final Storage zip(
+      String name, BiFunction<Object, Object, Object> function, Storage arg, boolean skipNa) {
     if (name != null && isOpVectorized(name)) {
       return runVectorizedZip(name, arg);
     }
@@ -188,7 +190,7 @@ public abstract class Storage {
     for (int i = 0; i < size(); i++) {
       Object it1 = getItemBoxed(i);
       Object it2 = i < arg.size() ? arg.getItemBoxed(i) : null;
-      if (it1 == null || it2 == null) {
+      if (skipNa && (it1 == null || it2 == null)) {
         builder.appendNoGrow(null);
       } else {
         builder.appendNoGrow(function.apply(it1, it2));
@@ -205,6 +207,24 @@ public abstract class Storage {
    */
   public Storage fillMissing(Object arg) {
     return fillMissingHelper(arg, new ObjectBuilder(size()));
+  }
+
+  /**
+   * Fills missing values in this storage, by using corresponding values from {@code other}.
+   *
+   * @param other the source of default values
+   * @return a new storage with missing values filled
+   */
+  public Storage fillMissingFrom(Storage other) {
+    var builder = new InferredBuilder(size());
+    for (int i = 0; i < size(); i++) {
+      if (isNa(i)) {
+        builder.appendNoGrow(other.getItemBoxed(i));
+      } else {
+        builder.appendNoGrow(getItemBoxed(i));
+      }
+    }
+    return builder.seal();
   }
 
   protected final Storage fillMissingHelper(Object arg, Builder builder) {
@@ -259,6 +279,23 @@ public abstract class Storage {
 
   public List<Object> toList() {
     return new StorageListView(this);
+  }
+
+  /**
+   * Counts the number of times each value has been seen before in this storage.
+   *
+   * @return a storage counting the number of times each value in this one has been seen before.
+   */
+  public Storage duplicateCount() {
+    long[] data = new long[size()];
+    HashMap<Object, Integer> occurenceCount = new HashMap<>();
+    for (int i = 0; i < size(); i++) {
+      var value = getItemBoxed(i);
+      var count = occurenceCount.getOrDefault(value, 0);
+      data[i] = count;
+      occurenceCount.put(value, count + 1);
+    }
+    return new LongStorage(data);
   }
 
   /**