+++ std-bits/table/src/main/java/org/enso/table/data/table/Column.java
@@ -171,4 +171,9 @@ public class Column {
   public Column slice(int offset, int limit) {
     return new Column(name, index.slice(offset, limit), storage.slice(offset, limit));
   }
+
+  /** @return a column counting value repetitions in this column. */
+  public Column duplicateCount() {
+    return new Column(name + "_duplicate_count", index, storage.duplicateCount());
+  }
 }